Adds a LongOperation dialog
The LongOperation Dialog is for when the plugins are in a block operation and the plugin needs to inform the user.
In this context, the BinanceSPOT plugin now waits for the WebSocket to be connected in a blocking manner. This fix an issue when the favorites list is filled at startup. When more than one symbol is in the list, not every symbol was signaled to received input because the WebSocket was not yet started, so the WatchList favorites was filled however not every symbol received its input

Fixes and improvements

Fixes the qDebug() line which sometimes cause a general exception failure when the UI was closed.

Sets the future value