Bug Fixes
All plugins are unloaded in the main application object closeEvent.
There are no guarantees that the plugins are unloaded before the objects destroyed in the destructor. And this situation may generate violations access when a plugin access, p.e. the logger. when has already been destroyed. So unloading the plugins before the destroyer can be a good handler.